requirements.txt
Django>=4.2,<5
djangorestframework>=3.16,<4
djangorestframework-simplejwt>=5.0,<6
django-cors-headers>=4.0,<5
python-dotenv>=1.0,<2
pycryptodome>=3.20,<4
gunicorn>=21.2,<22
whitenoise>=6.6,<7
dj-database-url>=2.2,<3
psycopg[binary]>=3.2,<4
